Protecting Video Routes and videoDetail, userDetail bug fix

+++ controllers/videoController.js
@@ -54,7 +54,6 @@ export const videoDetail = async (req, res) => {
   try {
     // populate : 객체 추출
     const video = await Video.findById(id).populate("creator");
-    console.log(video);
     res.render("videoDetail", { pageTitle: video.title, video });
   } catch (error) {
     res.redirect(routes.home);
@@ -68,8 +67,13 @@ export const getEditVideo = async (req, res) => {
 
   try {
     const video = await Video.findById(id);
-    res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    if (video.creator !== req.user.id) {
+      throw Error();
+    } else {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (error) {
+    console.log(error);
     res.redirect(routes.home);
   }
 };
@@ -94,7 +98,12 @@ export const deleteVideo = async (req, res) => {
   } = req;
 
   try {
-    await Video.findByIdAndRemove({ _id: id });
+    const video = await Video.findById(id);
+    if (video.creator !== req.user.id) {
+      throw Error();
+    } else {
+      await Video.findByIdAndRemove({ _id: id });
+    }
   } catch (error) {
     console.log(error);
   }
